+++ src/plonemeeting/restapi/services/search.py
@@ -3,20 +3,29 @@
 from imio.restapi.services.search import SearchGet
 from plone import api
 from plone.restapi.deserializer import boolean_value
-from plonemeeting.restapi.config import CONFIG_ID_ERROR
 from plonemeeting.restapi.config import CONFIG_ID_NOT_FOUND_ERROR
+from plonemeeting.restapi.config import TYPE_WITHOUT_CONFIG_ID_ERROR
 from plonemeeting.restapi.utils import check_in_name_of
 
 
 class PMSearchGet(SearchGet):
+    """Ease search of items and meetings related to a MeetingConfig.
+       If config_id in REQUEST, we will configure some shortcuts,
+       if not, the endpoint will behave as default."""
+
     def __init__(self, context, request):
         super(PMSearchGet, self).__init__(context, request)
         self.tool = api.portal.get_tool("portal_plonemeeting")
         config_id = self._config_id
         self.type = self._type
-        self.cfg = self.tool.get(config_id, None)
-        if not self.cfg:
-            raise Exception(CONFIG_ID_NOT_FOUND_ERROR % config_id)
+        # if config_id is given, config_id must exist
+        if config_id:
+            self.cfg = self.tool.get(config_id, None)
+            if not self.cfg:
+                raise Exception(CONFIG_ID_NOT_FOUND_ERROR % config_id)
+        # type without config_id is nonsense
+        if self.type and not config_id:
+            raise Exception(TYPE_WITHOUT_CONFIG_ID_ERROR % self.type)
 
     @property
     def _additional_fields(self):
@@ -25,13 +34,13 @@ class PMSearchGet(SearchGet):
 
     @property
     def _config_id(self):
-        if "config_id" not in self.request.form:
-            raise Exception(CONFIG_ID_ERROR)
-        return self.request.form.get("config_id")
+        return self.request.form.get("config_id", None)
 
     @property
     def _type(self):
-        if "type" not in self.request.form:
+        # when using "config_id", default searched type is "item"
+        if "type" not in self.request.form and \
+           "config_id" in self.request.form:
             self.request.form["type"] = "item"
         return self.request.form.get("type")
 
@@ -54,7 +63,8 @@ class PMSearchGet(SearchGet):
         form = self.request.form
 
         # config_id is actually the getConfigId index
-        query["getConfigId"] = form["config_id"]
+        if "config_id" in form:
+            query["getConfigId"] = form["config_id"]
 
         # extend batch? DEFAULT_BATCH_SIZE = 25
         # self.request.form['b_size'] = 50