+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-1.0b1 (unreleased)
+1.0b1 (2021-02-03)
 ------------------
 
 - Override default `PMBrainJSONSummarySerializer` for `ICatalogBrain` from